app/api-server-fastapi/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
PyJWT>=2.8.0
bcrypt>=4.1.2
pydantic>=2.5.3
aiosqlite>=0.19.0
aiomysql>=0.2.0
boto3>=1.34.0
redis>=5.0.1
